Refactor rollback: IRollbackManager, RollbackMgr, SQLite3 backend
Split monolithic rollback.cpp into shared RollbackMgr (server/rollback.cpp),
SQLite implementation (rollback_sqlite3.*), and existing rollback_interface.

Server selects rollback via world.mt rollback_backend (sqlite3 only in this PR).
PostgreSQL backend will follow in a separate change.

Preserves upstream addAction buffer ordering and getNodeActors statement order
per review feedback.

Made-with: Cursor
