+++ doc/lua_api.md
@@ -7504,6 +7504,9 @@ Item handling
 Rollback
 --------
 
+These functions require rollback recording to be enabled on the server (`enable_rollback_recording = true`).
+See [`world.mt` documentation](world_format.md) for selecting the rollback storage backend.
+
 * `core.rollback_get_node_actions(pos, range, seconds, limit)`:
   returns `{{actor, pos, time, oldnode, newnode}, ...}`
     * Find who has done something to a node, or near a node
