+++ src/rollback_interface.h
@@ -102,17 +102,17 @@ public:
 	virtual bool isActorGuess() = 0;
 	virtual void setActor(const std::string &actor, bool is_guess) = 0;
 	virtual std::string getSuspect(v3s16 p, float nearness_shortcut,
-	                               float min_nearness) = 0;
+		float min_nearness) = 0;
 
 	virtual ~IRollbackManager() = default;;
 	virtual void flush() = 0;
 	// Get all actors that did something to position p, but not further than
 	// <seconds> in history
 	virtual std::list<RollbackAction> getNodeActors(v3s16 pos, int range,
-	                time_t seconds, int limit) = 0;
+		time_t seconds, int limit) = 0;
 	// Get actions to revert <seconds> of history made by <actor>
 	virtual std::list<RollbackAction> getRevertActions(const std::string &actor,
-	                time_t seconds) = 0;
+		time_t seconds) = 0;
 };
